trying full script using base64 encoding on license key in the hopes the quotes will be handled correctly on azure's end

+++ datakeeper-cluster/Scripts/PrepareDataKeeperNodes.ps1
@@ -10,20 +10,26 @@ param
     [String] $AdminBase64Password,
 		
 		[Parameter(Mandatory=$true)]
-    [String] $TempLicense
+    [String] $TempLicense,
+		
+		[Parameter(Mandatory=$true)]
+    [int] $NodeIndex	
 )
 
 function TraceInfo($log)
 {
-    if ($script:LogFile -ne $null)
-    {
-        "$(Get-Date -format 'MM/dd/yyyy HH:mm:ss') $log" | Out-File -Confirm:$false -FilePath $script:LogFile -Append
-    }    
+     "$(Get-Date -format 'MM/dd/yyyy HH:mm:ss') $log" | Add-Content -Confirm:$false $logFile 
 }
 
 Set-StrictMode -Version 3
 $datetimestr = (Get-Date).ToString("yyyyMMddHHmmssfff")        
-$script:LogFile = "$env:windir\Temp\HpcPrepareCNLog-$datetimestr.txt"
+$logFile = "$env:windir\Temp\PrepareDataKeeperNode_log-$datetimestr.txt"
+
+$license = [System.Text.Encoding]::UTF8.GetString([System.Convert]::FromBase64String($TempLicense))
+TraceInfo "PrepareDataKeeperNodes script called with TempLicense:"
+TraceInfo $TempLicense
+TraceInfo "TempLicense converted to: "
+TraceInfo $license
 
 $AdminPassword = [System.Text.Encoding]::UTF8.GetString([System.Convert]::FromBase64String($AdminBase64Password))
 $domainNetBios = $DomainFQDN.Split(".")[0].ToUpper()
@@ -60,5 +66,64 @@ TraceInfo "Start to set initialize data disk(s)"
 Get-Disk | Where partitionstyle -eq 'raw' | Initialize-Disk -PartitionStyle MBR -PassThru |	New-Partition -AssignDriveLetter -UseMaximumSize | Format-Volume -FileSystem NTFS -Confirm:$false
 TraceInfo "Finished formatting data disk(s)"
 
-TraceInfo "Restart after 30 seconds"
-Start-Process -FilePath "cmd.exe" -ArgumentList "/c shutdown /r /t 30"
+TraceInfo "Installing DataKeeper license"
+$licfile = $env:windir\SysWOW64\LKLicense\extmirrsvc.lic
+Add-Content $licfile 'FEATURE dkwfce steeleye 1.000 03-nov-2015 uncounted HOSTID=ANY TS_OK SIGN="'
+Add-Content $licfile $license
+Add-Content $licfile '"'
+TraceInfo "Finished installing DataKeeper license"
+
+TraceInfo "Enabling and Configuring WSFC"
+Install-WindowsFeature -Name Failover-Clustering -IncludeManagementTools
+Import-Module FailoverClusters
+# TODO: add node to cluster 
+TraceInfo "Finished Configuring WSFC"
+
+TraceInfo "Verifying ExtMirr Service is Running"
+while(-Not $(Test-Path $env:windir\SysWOW64\LKLicense\extmirrsvc.lic)) {
+		Start-Sleep -Seconds 5
+}
+Restart-Service extmirrsvc
+Start-Sleep -Seconds 10
+if(-Not $(Get-Service extmirrsvc).Status -eq "Running") {
+	TraceInfo "ExtMirr Service failed to start. Mirror not created."
+	return 1
+}
+
+TraceInfo "Enabling remote script execution between nodes."
+Enable-PSRemoting -force
+winrm s winrm/config/client '@{TrustedHosts="sios-0,sios-1"}'
+
+Restart-Service winrm
+Start-Sleep -Seconds 10
+
+if(-Not $(Get-Service extmirrsvc).Status -eq "Running") {
+	TraceInfo "Remote Windows Management failed to start."
+	return 1
+}
+
+if($NodeIndex -eq 0) {	
+	TraceInfo "Waiting on sios-1 to start extmirrsvc"
+	while($true) {
+		if($(Get-Service extmirrsvc -ComputerName "sios-1").Status -eq "Running") {
+			break
+		} else {
+			Start-Sleep 10
+		}
+	}
+
+	TraceInfo "Creating mirror on volume F"
+	New-DataKeeperJob "Volume F" "Initial Mirror" "sios-0" "10.0.0.5" "F" "sios-1" "10.0.0.6" "F" "Async"
+	New-DataKeeperMirror "10.0.0.5" "F" "10.0.0.6" "F" "Async"
+	
+	# TODO: register mirror with cluster
+	
+	TraceInfo "Finished creating mirror. Restarting remote node (sios-1)."
+	Restart-Computer sios-1 -force
+	
+	TraceInfo "Restarting this node (sios-0) after 30 seconds"
+	Start-Process -FilePath "cmd.exe" -ArgumentList "/c shutdown /r /t 30"
+
+} else {
+	TraceInfo "Completed provisioning sios-1, waiting on sios-0 to restart this node."
+}
